requirements.txt
pydantic
ollama
duckduckgo-search
apscheduler
dateparser
pytz
slack-bolt
httpx
beautifulsoup4
python-dotenv
pytest
